refactor: 최종 우승자 로직 테스트 코드 위치 변경
